+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- biological RBE calculation from alpha and beta kernels
+- biological based optimization
+- alpha and beta parameters to dij with function `get_reference_lq_params` to get them for a given ct and cst
+- quantity resolver that checks for presence of quantities and instantiates the required ones
+
 ### Changed
 - resampling in BLD now uses interpolation to return mask on the grid provided by the dose engine
 - in SVDPB field_grid is now built before the resampling of the beamlet mask to guarantee matching grids